requirements.txt
flask
flask-cors
openai
python-dotenv
PyJWT
requests
gunicorn
# MCP packages - try one of these approaches:
mcp
# OR if the above doesn't work, try:
# git+https://github.com/modelcontextprotocol/python-sdk.git